Use correct time_ctrl for determining visibillity
