+++ crates/re_viewer_context/src/time_control.rs
@@ -1,6 +1,8 @@
 use std::collections::BTreeMap;
 
-use re_entity_db::{EditableAutoValue, TimeCounts, TimesPerTimeline};
+use re_entity_db::{
+    EditableAutoValue, EntityTree, TimeCounts, TimeHistogramPerTimeline, TimesPerTimeline,
+};
 use re_log_types::{Duration, TimeInt, TimeRange, TimeRangeF, TimeReal, TimeType, Timeline};
 
 use crate::NeedsRepaint;
@@ -539,6 +541,22 @@ impl TimeControl {
         if let Some(state) = self.states.get_mut(&self.timeline) {
             state.view = None;
         }
+    }
+
+    /// Returns whether the given tree has any data logged in the current timeline,
+    /// or has any timeless messages.
+    pub fn tree_has_data_in_current_timeline(&self, tree: &EntityTree) -> bool {
+        let top_time_histogram = &tree.subtree.time_histogram;
+        top_time_histogram.has_timeline(self.timeline())
+            || top_time_histogram.num_timeless_messages() > 0
+    }
+
+    /// Returns whether the given component has any data logged in the current timeline.
+    pub fn component_has_data_in_current_timeline(
+        &self,
+        component_stat: &TimeHistogramPerTimeline,
+    ) -> bool {
+        component_stat.has_timeline(self.timeline()) || component_stat.num_timeless_messages() > 0
     }
 }
