+++ crates/re_viewer_context/src/viewer_context.rs
@@ -1,7 +1,7 @@
 use ahash::HashMap;
 use parking_lot::RwLock;
 
-use re_entity_db::{entity_db::EntityDb, EntityTree, TimeHistogramPerTimeline};
+use re_entity_db::entity_db::EntityDb;
 
 use crate::{
     query_context::DataQueryResult, AppOptions, ApplicableEntities, ApplicationSelectionState,
@@ -83,23 +83,6 @@ impl<'a> ViewerContext<'a> {
     /// The current time query, based on the current time control.
     pub fn current_query(&self) -> re_data_store::LatestAtQuery {
         self.rec_cfg.time_ctrl.read().current_query()
-    }
-
-    /// Returns whether the given tree has any data logged in the current timeline,
-    /// or has any timeless messages.
-    pub fn tree_has_data_in_current_timeline(&self, tree: &EntityTree) -> bool {
-        let top_time_histogram = &tree.subtree.time_histogram;
-        top_time_histogram.has_timeline(self.rec_cfg.time_ctrl.read().timeline())
-            || top_time_histogram.num_timeless_messages() > 0
-    }
-
-    /// Returns whether the given component has any data logged in the current timeline.
-    pub fn component_has_data_in_current_timeline(
-        &self,
-        component_stat: &TimeHistogramPerTimeline,
-    ) -> bool {
-        component_stat.has_timeline(self.rec_cfg.time_ctrl.read().timeline())
-            || component_stat.num_timeless_messages() > 0
     }
 }
